+++ packages/next/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @chatbotkit/next
 
+## 1.26.1
+
+### Patch Changes
+
+- 3c96292: Fixed pnpm export.
+
 ## 1.21.7
 
 ### Patch Changes