+++ packages/nextauth/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @chatbotkit/nextauth
 
+## 1.26.2
+
+### Patch Changes
+
+- Updated dependencies [53fe36a]
+  - @chatbotkit/sdk@1.26.2
+
+## 1.26.1
+
+### Patch Changes
+
+- 3c96292: Fixed pnpm export.
+- Updated dependencies [3c96292]
+  - @chatbotkit/sdk@1.26.1
+
 ## 1.26.0
 
 ### Patch Changes