+++ packages/sdk/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @chatbotkit/sdk
 
+## 1.26.2
+
+### Patch Changes
+
+- 53fe36a: Updated types and new agentic methods.
+
+## 1.26.1
+
+### Patch Changes
+
+- 3c96292: Fixed pnpm export.
+- Updated dependencies [3c96292]
+  - @chatbotkit/fetch@1.26.1
+
 ## 1.26.0
 
 ### Minor Changes