+++ packages/widget/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @chatbotkit/widget
 
+## 1.26.1
+
+### Patch Changes
+
+- 3c96292: Fixed pnpm export.
+
 ## 1.25.0
 
 ### Minor Changes